add componentization and test scaffolding
